Updates: simplify includes, update test scripts, clearer naming of some elements
